// src/annotator/types/SerializedAnnotationType.ts
export type SerializedAnnotationType = {
  annotationBoundingBoxHeight: number;
  annotationBoundingBoxWidth: number;
  annotationBoundingBoxX: number;
  annotationBoundingBoxY: number;
  annotationCategoryColor: string;
  annotationCategoryId: string;
  annotationCategoryName: string;
  annotationId: string;
  annotationMask: string;
};